l10n_it_delivery_note: split common test class
