+++ l10n_it_delivery_note/tests/__init__.py
@@ -1,1 +1,2 @@
+from . import delivery_note_common
 from . import test_stock_delivery_note_invoicing
